split overloaded lists component into lists and tasks components
