+++ front-end/src/App.tsx
@@ -18,7 +18,7 @@ function App(): ReactElement {
   const defaultDark: boolean = window.matchMedia(
     "(prefers-color-scheme:dark)"
   ).matches;
-  const [theme, setTheme] = useLocalStorage(
+  const [theme, setTheme] = useLocalStorage<string>(
     "theme",
     defaultDark ? "dark" : "light"
   );
@@ -29,7 +29,7 @@ function App(): ReactElement {
    * the state as a User
    * If the userData is null, set the user in the state as null
    */
-  const userData = useLoaderData();
+  const userData: object | unknown = useLoaderData();
   const [user, setUser] = useState<User | null>(
     typeof userData === "object" && userData !== null
       ? (userData as User)
