+++ front-end/src/pages/Workflow.tsx
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { useOutletContext } from "react-router-dom";
-import Lists from "../components/workflow/Lists";
+import Lists from "../components/workflow/lists/Lists";
 import { ContextType } from "../types/userTypes";
 import Airports from "../components/workflow/Airports";
 import NamedLocations from "../components/workflow/NamedLocations";
@@ -16,7 +16,6 @@ const Workflow: React.FC = (): ReactElement => {
       <WeatherProvider>
         <div className="d-flex flex-column">
           <div className="d-flex flex-grow-1 align-items-center my-3">
-            {/* airports and locations stuff */}
             <div className="d-flex flex-column flex-grow-1 h-100 mx-1">
               <Airports user={user} setUser={setUser} theme={theme} />
             </div>
